Store-service. ISSUE-47. Пофиксил hoMuchAtTimeDTO

+++ store-service/src/main/java/vvp_company/storeservice/dto/nested/ItemResponseDTO.java
@@ -14,5 +14,5 @@ public class ItemResponseDTO {
     private ItemType itemType;
     private String itemName;
     private Integer itemWeight;
-    private Integer itemQuantity;
+    private Long itemQuantity;
 }
